Introduce applyRefactorings library function
