+++ apply-refact.cabal
@@ -38,6 +38,10 @@ library
                , syb
                , mtl
                , process
+               , transformers
+               , temporary
+               , filemanip
+               , unix-compat
                , directory
   hs-source-dirs:      src
   default-language:    Haskell2010
