+++ src/Main.hs
@@ -13,9 +13,7 @@ module Main where
 
 import Language.Haskell.GHC.ExactPrint
 import Language.Haskell.GHC.ExactPrint.Print
-import Language.Haskell.GHC.ExactPrint.Delta
 import Language.Haskell.GHC.ExactPrint.Parsers (parseModuleWithOptions)
-import Language.Haskell.GHC.ExactPrint.Types
 import Language.Haskell.GHC.ExactPrint.Utils
 
 
@@ -30,7 +28,6 @@ import Options.Applicative
 import Data.Maybe
 import Control.Monad.Trans.Maybe
 import Data.List hiding (find)
-import Data.Ord
 
 import System.IO
 import System.IO.Temp
@@ -51,7 +48,6 @@ import SrcLoc
 import Text.Read
 import Data.Char
 
-data Verbosity = Silent | Normal | Loud deriving (Eq, Show, Ord)
 
 parseVerbosity :: Monad m => String -> m Verbosity
 parseVerbosity s =
@@ -184,11 +180,6 @@ filterFilename = do
 
 -- Pipe
 
-refactOptions :: PrintOptions Identity String
-refactOptions = stringOptions { epRigidity = RigidLayout }
-
-rigidLayout :: DeltaOptions
-rigidLayout = deltaOptions RigidLayout
 
 runPipe :: Options -> FilePath  -> IO ()
 runPipe Options{..} file = do
@@ -228,51 +219,6 @@ runPipe Options{..} file = do
            Just f  -> do
             when (verb == Loud) (traceM $ "Writing result to " ++ f)
             writeFile f output
-
--- Filters out overlapping ideas, picking the first idea in a set of overlapping ideas.
--- If two ideas start in the exact same place, pick the largest edit.
-removeOverlap :: Verbosity -> [(String, [Refactoring R.SrcSpan])] -> [(String, [Refactoring R.SrcSpan])]
-removeOverlap verb = dropOverlapping . sortBy f . summarize
-  where
-    -- We want to consider all Refactorings of a single idea as a unit, so compute a summary
-    -- SrcSpan that encompasses all the Refactorings within each idea.
-    summarize :: [(String, [Refactoring R.SrcSpan])] -> [(String, (R.SrcSpan, [Refactoring R.SrcSpan]))]
-    summarize ideas = [ (s, (foldr1 summary (map pos rs), rs)) | (s, rs) <- ideas, not (null rs) ]
-
-    summary (R.SrcSpan sl1 sc1 el1 ec1)
-            (R.SrcSpan sl2 sc2 el2 ec2) =
-      let (sl, sc) = case compare sl1 sl2 of
-                      LT -> (sl1, sc1)
-                      EQ -> (sl1, min sc1 sc2)
-                      GT -> (sl2, sc2)
-          (el, ec) = case compare el1 el2 of
-                      LT -> (el2, ec2)
-                      EQ -> (el2, max ec1 ec2)
-                      GT -> (el1, ec1)
-      in R.SrcSpan sl sc el ec
-
-    -- Order by span start. If starting in same place, order by size.
-    f (_,(s1,_)) (_,(s2,_)) =
-      comparing startLine s1 s2 <> -- s1 first if it starts on earlier line
-      comparing startCol s1 s2 <>  --             or on earlier column
-      comparing endLine s2 s1 <>   -- they start in same place, s2 comes
-      comparing endCol s2 s1       -- first if it ends later
-      -- else, completely same span, so s1 will be first
-
-    dropOverlapping [] = []
-    dropOverlapping (p:ps) = go p ps
-    go (s,(_,rs)) [] = [(s,rs)]
-    go p@(s,(_,rs)) (x:xs)
-      | p `overlaps` x = (if verb > Silent
-                          then trace ("Ignoring " ++ show (snd (snd x)) ++ " due to overlap.")
-                          else id) go p xs
-      | otherwise = (s,rs) : go x xs
-    -- for overlaps, we know s1 always starts <= s2, due to our sort
-    overlaps (_,(s1,_)) (_,(s2,_)) =
-      case compare (startLine s2) (endLine s1) of
-        LT -> True
-        EQ -> startCol s2 <= endCol s1
-        GT -> False
 
 data LoopOption = LoopOption
                     { desc :: String
