+++ src/Refact/Apply.hs
@@ -3,29 +3,43 @@
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE NoMonomorphismRestriction #-}
-module Refact.Apply (runRefactoring)  where
+module Refact.Apply
+  (
+    runRefactoring
+  , applyRefactorings
+
+  -- * Support for runPipe in the main process
+  , Verbosity(..)
+  , rigidLayout
+  , removeOverlap
+  , refactOptions
+  )  where
 
 import Language.Haskell.GHC.ExactPrint
-import Language.Haskell.GHC.ExactPrint.Parsers
 import Language.Haskell.GHC.ExactPrint.Annotate
+import Language.Haskell.GHC.ExactPrint.Delta
+import Language.Haskell.GHC.ExactPrint.Parsers
+import Language.Haskell.GHC.ExactPrint.Print
 import Language.Haskell.GHC.ExactPrint.Types
 import Language.Haskell.GHC.ExactPrint.Utils
 
+import Data.Maybe
+import Data.List hiding (find)
+import Data.Ord
+
+import Control.Monad
+import Control.Monad.State
+import Control.Monad.Identity
 import Data.Data
 import Data.Generics.Schemes
 
 import HsExpr as GHC hiding (Stmt)
-import qualified HsBinds as GHC
-import qualified HsDecls as GHC
 import HsImpExp
-import qualified Module as GHC
 import HsSyn hiding (Pat, Stmt)
 import SrcLoc
-import qualified SrcLoc as GHC
-import qualified RdrName as GHC
+import qualified GHC hiding (parseModule)
 import qualified OccName as GHC
-import Data.Generics
-import Control.Monad.State
+import Data.Generics hiding (GT)
 
 import qualified Data.Map as Map
 
@@ -33,13 +47,91 @@ import System.IO.Unsafe
 
 import Control.Arrow
 
-import Data.Maybe
+import Debug.Trace
 
+import Data.Monoid
+import Refact.Fixity
 import Refact.Types hiding (SrcSpan)
 import qualified Refact.Types as R
 import Refact.Utils (Stmt, Pat, Name, Decl, M, Expr, Type
-                    , modifyAnnKey, replaceAnnKey, Import)
+                    , modifyAnnKey, replaceAnnKey, Import, toGhcSrcSpan)
 
+-- library access to perform the substitutions
+
+refactOptions :: PrintOptions Identity String
+refactOptions = stringOptions { epRigidity = RigidLayout }
+
+rigidLayout :: DeltaOptions
+rigidLayout = deltaOptions RigidLayout
+
+-- | Apply a set of refactorings as supplied by hlint
+applyRefactorings :: Maybe (Int, Int) -> [(String, [Refactoring R.SrcSpan])] -> FilePath -> IO String
+applyRefactorings optionsPos inp file = do
+  (as, m) <- either (error . show) (uncurry applyFixities)
+              <$> parseModuleWithOptions rigidLayout file
+  let noOverlapInp = removeOverlap Silent inp
+      refacts = (fmap . fmap . fmap) (toGhcSrcSpan file) <$> noOverlapInp
+
+      posFilter (_, rs) =
+        case optionsPos of
+          Nothing -> True
+          Just p  -> any (flip spans p . pos) rs
+      filtRefacts = filter posFilter refacts
+
+  -- need a check here to avoid overlap
+  (ares, res) <- return . flip evalState 0 $
+                          foldM (uncurry runRefactoring) (as, m) (concatMap snd filtRefacts)
+  let output = runIdentity $ exactPrintWithOptions refactOptions res ares
+  return output
+
+data Verbosity = Silent | Normal | Loud deriving (Eq, Show, Ord)
+
+-- Filters out overlapping ideas, picking the first idea in a set of overlapping ideas.
+-- If two ideas start in the exact same place, pick the largest edit.
+removeOverlap :: Verbosity -> [(String, [Refactoring R.SrcSpan])] -> [(String, [Refactoring R.SrcSpan])]
+removeOverlap verb = dropOverlapping . sortBy f . summarize
+  where
+    -- We want to consider all Refactorings of a single idea as a unit, so compute a summary
+    -- SrcSpan that encompasses all the Refactorings within each idea.
+    summarize :: [(String, [Refactoring R.SrcSpan])] -> [(String, (R.SrcSpan, [Refactoring R.SrcSpan]))]
+    summarize ideas = [ (s, (foldr1 summary (map pos rs), rs)) | (s, rs) <- ideas, not (null rs) ]
+
+    summary (R.SrcSpan sl1 sc1 el1 ec1)
+            (R.SrcSpan sl2 sc2 el2 ec2) =
+      let (sl, sc) = case compare sl1 sl2 of
+                      LT -> (sl1, sc1)
+                      EQ -> (sl1, min sc1 sc2)
+                      GT -> (sl2, sc2)
+          (el, ec) = case compare el1 el2 of
+                      LT -> (el2, ec2)
+                      EQ -> (el2, max ec1 ec2)
+                      GT -> (el1, ec1)
+      in R.SrcSpan sl sc el ec
+
+    -- Order by span start. If starting in same place, order by size.
+    f (_,(s1,_)) (_,(s2,_)) =
+      comparing startLine s1 s2 <> -- s1 first if it starts on earlier line
+      comparing startCol s1 s2 <>  --             or on earlier column
+      comparing endLine s2 s1 <>   -- they start in same place, s2 comes
+      comparing endCol s2 s1       -- first if it ends later
+      -- else, completely same span, so s1 will be first
+
+    dropOverlapping [] = []
+    dropOverlapping (p:ps) = go p ps
+    go (s,(_,rs)) [] = [(s,rs)]
+    go p@(s,(_,rs)) (x:xs)
+      | p `overlaps` x = (if verb > Silent
+                          then trace ("Ignoring " ++ show (snd (snd x)) ++ " due to overlap.")
+                          else id) go p xs
+      | otherwise = (s,rs) : go x xs
+    -- for overlaps, we know s1 always starts <= s2, due to our sort
+    overlaps (_,(s1,_)) (_,(s2,_)) =
+      case compare (startLine s2) (endLine s1) of
+        LT -> True
+        EQ -> startCol s2 <= endCol s1
+        GT -> False
+
+-- ---------------------------------------------------------------------
 
 -- Perform the substitutions
